requirements.txt
pdfplumber==0.10.3
pdf2image==3.1.0
Pillow==10.1.0